Revert "ItemData 오류 수정"
This reverts commit 297d1a6dbe3dac520ccc2066aa0716cd5bf91b0d.

+++ Assets/Scripts/UI/Inventory/ItemData/ItemData.cs
@@ -7,7 +7,7 @@ public abstract class ItemData
     public string SID => _sid;
     public string Name => _name;
     public string Tooltip => _tooltip;
-    public float Weight => _weight;
+    public int Weight => _weight;
     public Sprite IconSprite => _iconSprite;
     public Item_Info_Item Info => _info;
